--- packages/bridge-ui-v2/src/components/AmountInput/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AmountInput.svelte';
