+++ packages/bridge-ui-v2/src/components/Bridge/AmountInput/index.ts
@@ -1,0 +1,1 @@
+export { default as AmountInput } from './AmountInput.svelte';
